softdsnl-act7-main/softdsnl-act7-main/text_classification_api/requirements.txt
tensorflow
numpy
django
djangorestframework